+++ DropBear/Alerts/Alert.swift
@@ -21,19 +21,3 @@ public protocol AlertType {
 extension Alert: AlertType {
     public typealias AlertButtonType = Button
 }
-
-public protocol AlertButton: RawRepresentable {
-    func buttonElement(from alert: XCUIElement) -> XCUIElement
-}
-
-extension AlertButton where RawValue == String {
-    public func buttonElement(from alert: XCUIElement) -> XCUIElement {
-        return alert.buttons[rawValue]
-    }
-}
-
-extension AlertButton where RawValue == Int {
-    public func buttonElement(from alert: XCUIElement) -> XCUIElement {
-        return alert.buttons.element(boundBy: rawValue)
-    }
-}
