+++ DropBear/XCTest/XCUIElement+Assert.swift
@@ -11,7 +11,8 @@ extension XCUIElement {
         return assert([assertion] + rest, file: file, line: line)
     }
 
-    func assert(_ assertions: [ElementAssertion], file: StaticString = #file, line: UInt = #line) -> Self {
+    @discardableResult
+    public func assert(_ assertions: [ElementAssertion], file: StaticString = #file, line: UInt = #line) -> Self {
         for assertion in assertions {
             if !assertion.assertion(self) {
                 let customMessage = assertion.message.map({ ":\n\($0)" }) ?? ""
